git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@7602 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ GameSettings.h
@@ -136,7 +136,7 @@ enum
 
 typedef struct
 {
-	INT8				bLastSavedGameSlot;					// The last saved game number goes in here	
+	INT32				bLastSavedGameSlot;					// The last saved game number goes in here	
 
 	UINT8				ubMusicVolumeSetting;               // Volume Setting
 	UINT8				ubSoundEffectsVolume;               // Volume Setting
